Begin genotype refactor
